change to update the existing table2 instead
