--- lua/wikis/commons/Components/Table2/ColumnUtil.lua
@@ -1,150 +1,0 @@
----
--- @Liquipedia
--- page=Module:Components/Table2/ColumnUtil
---
--- Please see https://github.com/Liquipedia/Lua-Modules to contribute
---
-
-local Lua = require('Module:Lua')
-
-local Array = Lua.import('Module:Array')
-local Logic = Lua.import('Module:Logic')
-local MathUtil = Lua.import('Module:MathUtil')
-local Table = Lua.import('Module:Table')
-local WidgetUtil = Lua.import('Module:Widget/Util')
-
-local ColumnUtil = {}
-
----Gets the column index for this cell
----@param columnIndexProp integer|string? - explicit column index from props
----@param columnIndexContext integer|string? - implicit column index from context
----@return integer
-function ColumnUtil.getColumnIndex(columnIndexProp, columnIndexContext)
-	local index = MathUtil.toInteger(columnIndexProp) or MathUtil.toInteger(columnIndexContext)
-	return index or 1
-end
-
----Merges column definition properties with cell properties
----Cell props take precedence over column props
----Uses lazy merging: only creates merged table when column def has inheritable properties
----@param cellProps table
----@param columnDef table?
----@return table mergedProps
-function ColumnUtil.mergeProps(cellProps, columnDef)
-	if not columnDef then
-		return cellProps
-	end
-
-	local inheritableProps = {
-		'align', 'shrink', 'nowrap', 'width', 'minWidth', 'maxWidth',
-		'sortType', 'unsortable',
-	}
-
-	local hasInheritableProps = false
-	for _, prop in ipairs(inheritableProps) do
-		if columnDef[prop] ~= nil then
-			hasInheritableProps = true
-			break
-		end
-	end
-
-	local hasCss = columnDef.css ~= nil
-	local hasClasses = columnDef.classes ~= nil
-	local hasAttributes = columnDef.attributes ~= nil
-
-	if not hasInheritableProps and not hasCss and not hasClasses and not hasAttributes then
-		return cellProps
-	end
-
-	local merged = Table.copy(cellProps or {})
-
-	if hasInheritableProps then
-		Array.forEach(inheritableProps, function(prop)
-			if merged[prop] == nil and columnDef[prop] ~= nil then
-				merged[prop] = columnDef[prop]
-			end
-		end)
-	end
-
-	if hasCss then
-		merged.css = Table.merge(columnDef.css, merged.css or {})
-	end
-
-	if hasClasses then
-		merged.classes = WidgetUtil.collect(columnDef.classes, merged.classes)
-	end
-
-	if hasAttributes then
-		merged.attributes = Table.merge(columnDef.attributes, merged.attributes or {})
-	end
-
-	return merged
-end
-
----Builds CSS rules for sizing
----@param width string?
----@param minWidth string?
----@param maxWidth string?
----@param existingCss table?
----@return table css
-function ColumnUtil.buildCss(width, minWidth, maxWidth, existingCss)
-	if not width and not minWidth and not maxWidth then
-		return existingCss or {}
-	end
-
-	return Table.merge(existingCss or {}, {
-		width = width,
-		['min-width'] = minWidth,
-		['max-width'] = maxWidth,
-	})
-end
-
----Builds data attributes for cell styling (align, nowrap, shrink)
----@param align string?
----@param nowrap (string|number|boolean)?
----@param shrink (string|number|boolean)?
----@param existingAttributes table?
----@return table attributes
-function ColumnUtil.buildCellAttributes(align, nowrap, shrink, existingAttributes)
-	local attrs = existingAttributes or {}
-
-	if align == 'right' or align == 'center' then
-		attrs['data-align'] = align
-	end
-
-	if Logic.readBool(nowrap) then
-		attrs['data-nowrap'] = ''
-	end
-
-	if Logic.readBool(shrink) then
-		attrs['data-shrink'] = ''
-	end
-
-	return attrs
-end
-
----Builds HTML attributes for cells and headers
----@param mergedProps table - merged cell/header properties
----@param additionalAttributeBuilders {[string]: function}? - optional callbacks for additional attributes
----@return table attributes
-function ColumnUtil.buildAttributes(mergedProps, additionalAttributeBuilders)
-	local attributes = mergedProps.attributes or {}
-
-	if mergedProps.colspan then
-		attributes.colspan = MathUtil.toInteger(mergedProps.colspan) or mergedProps.colspan
-	end
-
-	if mergedProps.rowspan then
-		attributes.rowspan = MathUtil.toInteger(mergedProps.rowspan) or mergedProps.rowspan
-	end
-
-	if additionalAttributeBuilders then
-		Table.iter.forEachPair(additionalAttributeBuilders, function(key, builder)
-			builder(attributes, mergedProps)
-		end)
-	end
-
-	return attributes
-end
-
-return ColumnUtil
